+++ apps/block_scout_web/lib/block_scout_web/chain.ex
@@ -38,11 +38,10 @@ defmodule BlockScoutWeb.Chain do
     Transaction,
     Transaction.StateChange,
     UserOperation,
-    Wei,
-    Withdrawal
+    Wei
   }
 
-  alias Explorer.Chain.Zkevm.TransactionBatch
+  alias Explorer.Chain.PolygonZkevm.TransactionBatch
   alias Explorer.PagingOptions
 
   defimpl Poison.Encoder, for: Decimal do
@@ -612,7 +611,7 @@ defmodule BlockScoutWeb.Chain do
     }
   end
 
-  defp paging_params(%Withdrawal{index: index}) do
+  defp paging_params(%{index: index}) do
     %{"index" => index}
   end