+++ apps/explorer/lib/explorer/chain/import/runner/polygon_zkevm/bridge_l1_tokens.ex
@@ -1,32 +1,31 @@
-defmodule Explorer.Chain.Import.Runner.Zkevm.LifecycleTransactions do
+defmodule Explorer.Chain.Import.Runner.PolygonZkevm.BridgeL1Tokens do
   @moduledoc """
-  Bulk imports `t:Explorer.Chain.Zkevm.LifecycleTransaction.t/0`.
+  Bulk imports `t:Explorer.Chain.PolygonZkevm.BridgeL1Token.t/0`.
   """
 
   require Ecto.Query
 
+  import Ecto.Query, only: [from: 2]
+
   alias Ecto.{Changeset, Multi, Repo}
   alias Explorer.Chain.Import
-  alias Explorer.Chain.Zkevm.LifecycleTransaction
+  alias Explorer.Chain.PolygonZkevm.BridgeL1Token
   alias Explorer.Prometheus.Instrumenter
-
-  import Ecto.Query, only: [from: 2]
 
   @behaviour Import.Runner
 
   # milliseconds
   @timeout 60_000
 
-  @type imported :: [LifecycleTransaction.t()]
+  @type imported :: [BridgeL1Token.t()]
 
   @impl Import.Runner
-  def ecto_schema_module, do: LifecycleTransaction
+  def ecto_schema_module, do: BridgeL1Token
 
   @impl Import.Runner
-  def option_key, do: :zkevm_lifecycle_transactions
+  def option_key, do: :polygon_zkevm_bridge_l1_tokens
 
   @impl Import.Runner
-  @spec imported_table_row() :: %{:value_description => binary(), :value_type => binary()}
   def imported_table_row do
     %{
       value_type: "[#{ecto_schema_module()}.t()]",
@@ -35,7 +34,6 @@ defmodule Explorer.Chain.Import.Runner.Zkevm.LifecycleTransactions do
   end
 
   @impl Import.Runner
-  @spec run(Multi.t(), list(), map()) :: Multi.t()
   def run(multi, changes_list, %{timestamps: timestamps} = options) do
     insert_options =
       options
@@ -44,12 +42,12 @@ defmodule Explorer.Chain.Import.Runner.Zkevm.LifecycleTransactions do
       |> Map.put_new(:timeout, @timeout)
       |> Map.put(:timestamps, timestamps)
 
-    Multi.run(multi, :insert_zkevm_lifecycle_transactions, fn repo, _ ->
+    Multi.run(multi, :insert_polygon_zkevm_bridge_l1_tokens, fn repo, _ ->
       Instrumenter.block_import_stage_runner(
         fn -> insert(repo, changes_list, insert_options) end,
         :block_referencing,
-        :zkevm_lifecycle_transactions,
-        :zkevm_lifecycle_transactions
+        :polygon_zkevm_bridge_l1_tokens,
+        :polygon_zkevm_bridge_l1_tokens
       )
     end)
   end
@@ -58,24 +56,24 @@ defmodule Explorer.Chain.Import.Runner.Zkevm.LifecycleTransactions do
   def timeout, do: @timeout
 
   @spec insert(Repo.t(), [map()], %{required(:timeout) => timeout(), required(:timestamps) => Import.timestamps()}) ::
-          {:ok, [LifecycleTransaction.t()]}
+          {:ok, [BridgeL1Token.t()]}
           | {:error, [Changeset.t()]}
   def insert(repo, changes_list, %{timeout: timeout, timestamps: timestamps} = options) when is_list(changes_list) do
     on_conflict = Map.get_lazy(options, :on_conflict, &default_on_conflict/0)
 
-    # Enforce Zkevm.LifecycleTransaction ShareLocks order (see docs: sharelock.md)
-    ordered_changes_list = Enum.sort_by(changes_list, & &1.id)
+    # Enforce BridgeL1Token ShareLocks order (see docs: sharelock.md)
+    ordered_changes_list = Enum.sort_by(changes_list, &{&1.address})
 
     {:ok, inserted} =
       Import.insert_changes_list(
         repo,
         ordered_changes_list,
-        for: LifecycleTransaction,
+        conflict_target: :address,
+        on_conflict: on_conflict,
+        for: BridgeL1Token,
         returning: true,
         timeout: timeout,
-        timestamps: timestamps,
-        conflict_target: :hash,
-        on_conflict: on_conflict
+        timestamps: timestamps
       )
 
     {:ok, inserted}
@@ -83,20 +81,20 @@ defmodule Explorer.Chain.Import.Runner.Zkevm.LifecycleTransactions do
 
   defp default_on_conflict do
     from(
-      tx in LifecycleTransaction,
+      t in BridgeL1Token,
       update: [
         set: [
-          # don't update `id` as it is a primary key 
-          # don't update `hash` as it is a unique index and used for the conflict target
-          is_verify: fragment("EXCLUDED.is_verify"),
-          inserted_at: fragment("LEAST(?, EXCLUDED.inserted_at)", tx.inserted_at),
-          updated_at: fragment("GREATEST(?, EXCLUDED.updated_at)", tx.updated_at)
+          decimals: fragment("EXCLUDED.decimals"),
+          symbol: fragment("EXCLUDED.symbol"),
+          inserted_at: fragment("LEAST(?, EXCLUDED.inserted_at)", t.inserted_at),
+          updated_at: fragment("GREATEST(?, EXCLUDED.updated_at)", t.updated_at)
         ]
       ],
       where:
         fragment(
-          "(EXCLUDED.is_verify) IS DISTINCT FROM (?)",
-          tx.is_verify
+          "(EXCLUDED.decimals, EXCLUDED.symbol) IS DISTINCT FROM (?, ?)",
+          t.decimals,
+          t.symbol
         )
     )
   end